fayvad_geosolutions/requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-browser-reload==1.12.1

# Testing
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.9.1
isort==5.12.0

# Documentation
Sphinx==7.2.6

# Development server
django-livereload-server==0.4